+++ R/SingleGroup-methods.R
@@ -775,7 +775,7 @@ setMethod(
 #'     \item{\code{'scorecontour'}}{expected total score contour plot}
 #'   }
 #'
-#'   Note that if \code{empiricalhist = TRUE} was used in estimation then
+#'   Note that if \code{dentype = 'empiricalhist'} was used in estimation then
 #'   the type \code{'empiricalhist'}
 #'   also will be available to generate the empirical histogram plot
 #' @param degrees numeric value ranging from 0 to 90 used in \code{plot} to compute angle